wrapText refactor

- [x] Refactoring for better readability
- [x] Check and update tests (only if is necessary)
- [x] Improve performance

___

About the tests, i think that is necessary redefine them (like i did on the original PR), for example:

**current test:**
Test wrapText › When text contain new lines › should respect new lines and break all words when width of each word is less than container width
**for this input:**
```
Hello
whats up
```
**and this width:**
`90 (80 without padding)`
**the current expected result is:**
```
Hello
whats
up
```
**But, if each letter has width of 10, then:**
```
Hello = 50
whats = 50
 (space) = 10
up = 20
```
**then:**
`whats + (space) + up = 50 + 10 + 20 = 80`
`80 <= 80 = true!`
**the correct result must be:**
```
Hello
whats up
```
(that is my result)