@ignacio-cuadra 2023-03-03
Performance log
```
{
  text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, \n' +
    'sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. \n' +
    'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
  maxWidth: 10,
  loops: 50000
}
```
old wrap text: 975.995ms (before PR)
new wrap text: 6.351s (before last commit)
optimized wrap text: 1.777s (after last commit)

(this is a reminder note)