@ignacio-cuadra 2023-03-03
```
{
  text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, \n' +
    'sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. \n' +
    'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
  maxWidth: 10,
  loops: 50000
}
```
optimized wrap text (wip): 1.110s (second optimization)
old wrap text: 914.679ms (before PR)
optimized wrap text v1: 1.638s (first optimization)
new wrap text: 6.321s (before optimization)