@ignacio-cuadra 2023-03-03
```
{
  text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, \n' +
    'sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. \n' +
    'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
  maxWidth: 10,
  loops: 50000
}
```
master: 966ms | getLineWidth calls 2,050,023
third optimization (current): 1,141ms | getLineWidth calls 7,500,000
second optimization: 1,167ms | getLineWidth calls 21,700,000
first optimization: 1,959ms | getLineWidth calls 27,900,000
before optimization: 6,201ms | getLineWidth calls 27,900,000