@ad1992 2023-03-14
@ignacio-cuadra Sorry about the delay, I have been busy making some changes to the way the text wrapping works to make it more stable, also adding more tests, and monitoring to make sure all is good. I still have a few open PRs which I will be merging very soon (some of them are touching the wrap algo as well).

As this PR is mainly refactoring so it can wait (Once text containers are completely stable we can come back to this but for now it would be better not to introduce any code change which is not necessary atm)

But since you were working on a bug fix about fixing the spaces thing, I wanted to ask you if you could do it above the current master code ?
If that can be done, then we can merge that fix and It would help us release the fix soon.