@ad1992 2023-03-14
> Yes, I can do it but that feature removes all trailing spaces (I mention it because it would be necessary to remove them in the tests).

Yep that's what we ultimately want and that would be anyways needed to update in the test so its completely fine