@ignacio-cuadra 2023-03-15
Oops, i'm wrong. I meant that this feature doesn't remove any trailing space (They are removed when the text is drawing).