@ad1992 2023-03-15
> Oops, i'm wrong. I meant that this feature doesn't remove any trailing space (They are removed when the text is drawing).

That would be when you fix the bug of spaces right? which you would be doing in a new PR as discussed on top of the master's code?