@ad1992 2023-04-19
@ignacio-cuadra been long time so wondering if you are working on the trailing space fix or should I pick it up ?