@Mrazator 2024-10-16
Just noticed this! Looks much better, though we went in the end for a full rewrite as part of CJK support. Hence closing in favor of https://github.com/excalidraw/excalidraw/pull/8530.